Module type Versions.VERSION_1_0

module type RNG = sig ... end

Random bytes generator

module type S = sig ... end

Nano ID generator

val pseudo_seeded : string -> (module RNG)

pseudo_seeded seed is a RNG module that generates pseudo random numbers based on the given seed. It is subject to the same limitations as Cryptokit.Random.pseudo_rng

module Make : functor (Rng : RNG) -> S

Make (Rng) is a Nano ID generator based on random number generator Rng

module Simple : functor () S

Simple () is a Nano ID generator based on the pseudo_seeded random number generator seeded with the current system time as given by Unix.gettimeofday.